parallel learning sonic 3 CPU threads
This is a parallel learning python code that is learning on the video game sonic. The games has to be bought on steam then importanted though running the openai python steam import code. This code runs with 3 threads and normally completes the first level of sonic in under a minute. 

Watch the video of it working here youtube: https://www.youtube.com/watch?v=oKpx9jT7GJk